requirements.txt
pandas>=2.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
